Tplot

tplot is a versatile plotting utility that handles various time series formats including vectors, matrices, functions, and strings (product IDs). It renders data as line plots, series plots, heatmaps, or spectrograms.

tplot offers flexible visualization options, allowing you to display multiple time series either across separate panels or overlaid within the same panel.

tplot seamlessly integrates with Speasy.jl, automatically downloading and converting data to DimArray when given a product ID string.

Built on Makie, tplot provides both interactive exploration capabilities and publication-quality output. It features dynamic data loading during zoom/pan operations, efficiently retrieving and rendering data on demand.

SpaceTools.tplot — Function

tplot(f, tas; legend=(; position=Right()), link_xaxes=true, rowgap=5, kwargs...)

Lay out multiple time series across different panels (rows) on one Figure / GridPosition f

If legend is nothing, no legend will be added to the plot. Otherwise, legend can be a NamedTuple containing options for legend placement and styling.

SpaceTools.tplot_panel — Function

Setup the panel on a position and plot multiple time series on it

Setup the panel on a position and plot multiple time series on it

tplot_panel(gp, ta::AbstractDimMatrix)

Plot a multivariate time series / spectrogram on a panel

tplot_panel(gp, ta::AbstractDimVector)

Plot a univariate time series on a panel. Only add legend when the axis contains multiple labels.

Interactive tplot of a function over a time range

tplot_panel(gp, ta, args...; kwargs...)

Extension interface for plotting custom data types. To support a new data type:

Define a method for get_data(ta, args...; kwargs...) that converts your type to a DimensionalData array

Optionally include metadata for labels, units, and other plotting attributes

SpaceTools.tplot_panel! — Function

Plot heatmap / overlay multiple columns of a time series on the same axis

tplot_panel!(ax::Axis, ta, args...; kwargs...)

Extension interface for plotting custom data types. See tplot_panel for more details.

Function as tplot argument for interactive exploration

tplot can handle functions that accept time intervals as arguments. This allows for creating interactive plots where data is dynamically fetched. So instead of the two-step process:

Fetch data: da = f(t0, t1)

Plot data: tplot(da)

We can combine these steps into a single command:

tplot(f, t0, t1)

This approach enables efficient interactive exploration of time series.

Note

For real-time interactivity, consider using the GLMakie backend instead of CairoMakie although it is possible to use tlims! or xlims! to update the plot dynamically.

Related packages

PyTplot

InteractiveViz.jl

SciQLop : A python application built on top of Qt to explore multivariate time series effortlessly,