Tplot

SpaceTools.tplot — Function

Setup the axis on a position and plot multiple time series on it

Plot a multivariate time series on a position in a figure

Interactive tplot of a function over a time range

Lay out multiple time series on the same figure across different panels (rows)

SpaceTools.tplot! — Function

tplot!(ax, tas; kwargs...)

Overlay multiple time series on the same axis

Overlay multiple columns of a time series on the same axis

References

PyTplot

InteractiveViz.jl