API Reference

SpaceTools.fac_matrix_make — Method

fac_matrix_make(vec::AbstractVector; xref=[1.0, 0.0, 0.0])

Generates a field-aligned coordinate (FAC) transformation matrix for a vector.

Arguments

vec: A 3-element vector representing the magnetic field

SpaceTools.tplot! — Method

tplot!(ax, tas; kwargs...)

Overlay multiple time series on the same axis

SpaceTools.tplot! — Method

Overlay multiple columns of a time series on the same axis

SpaceTools.tplot! — Method

Overlay multiple columns of a time series on the same axis

SpaceTools.tplot — Method

Lay out multiple time series on the same figure across different panels (rows)

SpaceTools.tplot — Method

Setup the axis on a position and plot multiple time series on it

SpaceTools.tplot — Method

Plot a multivariate time series on a position in a figure

```