API Reference

DSP.Filters.filtfilt

SpaceTools.add_legend!

SpaceTools.amap

SpaceTools.check_mva_mat

SpaceTools.fac_mat

SpaceTools.mva

SpaceTools.mva_mat

SpaceTools.plot_attributes

SpaceTools.pspectrum

SpaceTools.rotate

SpaceTools.smooth

SpaceTools.tplot

SpaceTools.tplot

SpaceTools.tplot

SpaceTools.tplot

SpaceTools.tplot

SpaceTools.tplot!

SpaceTools.tplot!

SpaceTools.tplot!

SpaceTools.tsheat

DSP.Filters.filtfilt — Method

References

https://docs.juliadsp.org/stable/filters/

SpaceTools.add_legend! — Method

Only add legend when the axis contains multiple labels

SpaceTools.amap — Method

amap for intersection math

https://github.com/rafaqz/DimensionalData.jl/issues/914

SpaceTools.check_mva_mat — Method

check_mva_mat(F; r=5, verbose=false)

Check the quality of the MVA result.

If λ₁ ≥ λ₂ ≥ λ₃ are 3 eigenvalues of the constructed matrix M, then a good indicator of nice fitting LMN coordinate system should have λ₂ / λ₃ > r.

SpaceTools.fac_mat — Method

fac_mat(vec::AbstractVector; xref=[1.0, 0.0, 0.0])

Generates a field-aligned coordinate (FAC) transformation matrix for a vector.

Arguments

vec: A 3-element vector representing the magnetic field

SpaceTools.mva — Method

mva(V::AbstractMatrix, B::AbstractMatrix; kwargs...)

Rotate a timeseries V into the LMN coordinates based on the reference field B.

Arguments

V::AbstractMatrix: The timeseries data to be transformed, where each column represents a component

B::AbstractMatrix: The reference field used to determine the minimum variance directions, where each column represents a component

See also: mva(Bx, By, Bz), rotate

SpaceTools.mva_mat — Method

mva_mat(Bx, By, Bz; verbose=false)

Generates a LMN coordinate transformation matrix from 3 orthogonal vectors Bx, By, Bz.

Perform minimum variance analysis to vector components defined in orthogonal coordinates Bx, By and Bz. Set check=true to check the reliability of the result.

The kth eigenvector can be obtained from the slice F.vectors[:, k].

SpaceTools.plot_attributes — Method

plot_attributes(ta)

Plot attributes for a time array

SpaceTools.pspectrum — Method

pspectrum(x::AbstractDimArray, spec::Spectrogram)
pspectrum(x::AbstractDimArray; nfft=256, noverlap=128, window=hamming)

Compute the power spectrum (time-frequency representation) of a time series using the short-time Fourier transform.

Returns a DimArray with frequency and original time dimensions.

See also: DSP.Spectrogram, DSP.stft

Reference

Matlab

SpaceTools.rotate — Method

rotate(ts::AbstractMatrix, mat::AbstractMatrix)

Coordinate-aware transformation of vector/matrix by rotation matrix(s) mat(s). Assume ts is a matrix of shape (n, 3).

SpaceTools.smooth — Method

smooth(da::AbstractDimArray, window; dim=Ti, suffix="_smoothed", kwargs...)

Smooths a time series by computing a moving average over a sliding window.

The size of the sliding window can be either:

Quantity: A time duration that will be converted to number of samples based on data resolution

Integer: Number of samples directly

Arguments

dims=Ti: Dimension along which to perform smoothing (default: time dimension)

suffix="_smoothed": Suffix to append to the variable name in output

kwargs...: Additional arguments passed to RollingWindowArrays.rolling

SpaceTools.tplot! — Method

tplot!(ax, tas; kwargs...)

Overlay multiple time series on the same axis

SpaceTools.tplot! — Method

Overlay multiple columns of a time series on the same axis

SpaceTools.tplot — Method

Interactive tplot of a function over a time range

SpaceTools.tplot — Method

Lay out multiple time series on the same figure across different panels (rows)

SpaceTools.tplot — Method

Setup the axis on a position and plot multiple time series on it

SpaceTools.tplot — Method

Plot a multivariate time series on a position in a figure

SpaceTools.tsheat — Method

tsheat(data; kwargs...)

Heatmap with better default attributes for time series.

References:

https://docs.makie.org/stable/reference/plots/heatmap